backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pika==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
stripe==7.8.0
sqlalchemy==2.0.23
alembic==1.13.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
slowapi==0.1.9
pyotp==2.9.0
qrcode==7.4.2
itsdangerous==2.1.2
cryptography==41.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
sentry-sdk==1.39.1
prometheus-client==0.19.0
